feat(cli): add cli separation binary
