+++ .gitignore
@@ -1,2 +1,5 @@
-/target/
+target
 **/*.rs.bk
+Cargo.lock
+!spider_cli/Cargo.lock
+.env
