+++ examples/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "spider_examples"
+version = "1.4.0"
+authors = ["madeindjs <contact@rousseau-alexandre.fr>"]
+description = "Multithreaded Web spider crawler written in Rust."
+repository = "https://github.com/madeindjs/spider"
+readme = "README.md"
+keywords = ["crawler", "spider"]
+categories = ["web-programming"]
+license = "MIT"
+documentation = "https://docs.rs/spider"
+
+[badges]
+maintenance = { status = "as-is" }
+
+[dependencies.spider]
+version = "1.4.0"
+path = "../spider"
+default-features = false
+
+[[example]]
+name = "example"
+path = "example.rs"
