+++ spider/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "spider"
+version = "1.4.1"
+authors = ["madeindjs <contact@rousseau-alexandre.fr>"]
+description = "Multithreaded Web spider crawler written in Rust."
+repository = "https://github.com/madeindjs/spider"
+readme = "README.md"
+keywords = ["crawler", "spider"]
+categories = ["web-programming"]
+license = "MIT"
+documentation = "https://docs.rs/spider"
+
+[badges]
+maintenance = { status = "as-is" }
+
+[dependencies]
+reqwest = { version = "0.11", features = ["blocking"] }
+scraper = "0.12"
+robotparser = "0.10"
+url = "2.2"
+rayon = "1.1"
+num_cpus = "1.13.0"
