+++ spider_cli/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "spider_cli"
+version = "1.4.0"
+authors = ["madeindjs <contact@rousseau-alexandre.fr>"]
+description = "Multithreaded Web spider crawler written in Rust."
+repository = "https://github.com/madeindjs/spider"
+readme = "README.md"
+keywords = ["crawler", "spider", "spider_cli"]
+categories = ["web-programming"]
+license = "MIT"
+documentation = "https://docs.rs/spider"
+include = ["src/*", "LICENSE", "README.md"]
+
+[badges]
+maintenance = { status = "as-is" }
+
+[dependencies.spider]
+version = "1.4.0"
+path = "../spider"
+default-features = false
+
+[[bin]]
+name = "spider"
+path = "src/main.rs"
+doc = false
