--- src/lib.rs
@@ -1,13 +1,0 @@
-extern crate rayon;
-extern crate reqwest;
-extern crate robotparser;
-extern crate scraper;
-extern crate url;
-extern crate num_cpus;
-
-/// Configuration structure for `Website`
-pub mod configuration;
-/// A page scraped
-pub mod page;
-/// A website to crawl
-pub mod website;
